MFC r319814, r319815, r319818:
Add a set of constants describing the ways a MAC and PHY can be connected.
While the initial need for this is to help support phy drivers which are
configured with FDT data, there is nothing devicetree-specific about the
concept or the names, so they are available for use even on non-FDT systems.

Add some utility functions to help a PHY driver on an FDT-configured
system retrieve its config data from the fdt data.

Convert if_ffec from local code and constants for mac<->phy connection type
to new common fdt helper code.